Switch to function references in API
If I pass a function directly, I must borrow it to pass it in a
recursive call. This can end in an infinite depth recursion in the rustc
typechecker, causing compilation to fail. Instead, I pass functions by
reference, so their borrow level remains constant throughout the
recursion, avoiding this issue.