Move to interior mutability for Data
Our tuplestore needs to take objects which refrence Data, so we need
them to be able to get a mutating handle to it. This change allows Data
to be shallow-clonable, at the cost of a bit of unsafety around Views.

This also lays the groundwork for Data being used in a multithreaded
context via the usage guards.

* Data can be cloned
* When you receive a View, you must keep the View around until you are
done with its fields. I may enforce this with accessor functions and
private fields at a later date.