Add a timeout option to running the datalog engine
Currently, I can only terminate the engine between rules. This is
because the engine is designed as a single threaded mutable data
structure, which means that if I just hard stopped it, invariants might
not be maintained.

Harder (mid-join or mid function execution) termination would be
best accomplished by moving to a multithreaded model and killing worker
threads.