Skip "crate" in auto-named inputs
In Rust 2015, I used ::foo in the EDSL to specify something relative to
the root of the crate, so ::foo::bar would create FooBarIn and FooBarOut
structs. In Rust 2018, the root of the crate is written crate::foo::bar.
In order to avoid having structs named CrateFooBarIn, we drop the
crate:: prefix when we see it.